pyproject.toml
[project]
name = "pulse-gen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "duckdb>=1.4.3",
    "google-play-scraper>=1.2.7",
    "langchain>=1.2.1",
    "langchain-openai>=1.1.6",
    "langgraph>=1.0.5",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "tqdm>=4.67.1",
]
